Use polymorphism for product handling

+++ kata1/src/main/java/org/example/ReceiptLineItem.java
@@ -3,9 +3,9 @@ package org.example;
 public class ReceiptLineItem {
     public String productName;
     public double productTotal;
-    public int productAmount;
+    public double productAmount;
 
-    public ReceiptLineItem(String productName, double productTotal, int productAmount) {
+    public ReceiptLineItem(String productName, double productTotal, double productAmount) {
         this.productName = productName;
         this.productTotal = productTotal;
         this.productAmount = productAmount;
